Fix the analyzer's title split and its clashing stringify macro
Two defects in the analyzer frontend, both surfaced by warnings the
Windows build had never counted.

The plot title is split at the first newline using a byte offset held in
an int, which is then compared against the string length as a signed
value and used for the remainder's length. Hold the offset in the type
the string functions actually return and convert once, at the drawing
call that wants its own signed length.

Separately, the version-string stringify macro redefined a macro of the
same name that the version header already provides with a different
meaning, so which of the two the rest of the file saw depended on include
order. Use the existing one instead of declaring a second.

+++ lame/frontend/gpkplotting.c
@@ -198,7 +198,7 @@ gpk_bargraph_draw(GtkWidget * widget, /* plot on this widged */
     gint16  width, height, x, y, barheight;
     GdkFont *fixed_font;
     GdkGC  *gc;
-    int     titleSplit;
+    size_t  titleSplit;
 
 
     gc = gdk_gc_new(widget->window);
@@ -225,18 +225,19 @@ gpk_bargraph_draw(GtkWidget * widget, /* plot on this widged */
             if (titleSplit && (titleSplit != strlen(title))) {
                 gdk_draw_text(*ppixmap, fixed_font,
                               widget->style->fg_gc[GTK_WIDGET_STATE(widget)],
-                              0, 10, title, titleSplit);
+                              0, 10, title, (gint) titleSplit);
 
                 gdk_draw_text(*ppixmap, fixed_font,
                               widget->style->fg_gc[GTK_WIDGET_STATE(widget)],
-                              0, 22, title + titleSplit + 1, (strlen(title) - titleSplit) - 1);
+                              0, 22, title + titleSplit + 1,
+                              (gint) (strlen(title) - titleSplit - 1));
 
 
             }
             else {
                 gdk_draw_text(*ppixmap, fixed_font,
                               widget->style->fg_gc[GTK_WIDGET_STATE(widget)],
-                              0, 10, title, strlen(title));
+                              0, 10, title, (gint) strlen(title));
             }
         }
